@@ -1,12 +1,13 @@
 apply plugin: 'com.android.application'
 android {
-    compileSdkVersion 18
-    buildToolsVersion "23.0.0"
+    compileSdkVersion 29
 
     defaultConfig {
         applicationId "com.nordicsemi.nrfUARTv2"
         minSdkVersion 18
-        targetSdkVersion 18
+        targetSdkVersion 29
+        versionCode 44
+        versionName "2.0(44)"
     }
 
 
@@ -19,5 +20,5 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:support-v4:23.0.0'
+    implementation 'androidx.localbroadcastmanager:localbroadcastmanager:1.0.0'
 }
